that was stupid, so make it less so
